Make 2nd and beyond items in Multi-Item Lots give correct items, added messaging for replacement process
Fixes the issue where places like FS Talisman+Morningstar, or BT Crimson set (+tin banishment catalyst), the 2nd+ items in a set would just be multiple of the first. This was because of multiple ItemLotParams sharing the same "getItemFlagId".

Reworked the "item matching" logic to expect multiple item lots with the same flag id. In Helpers.OverwriteItemLots(), keep track of how many items for a specific flag we've placed or tried to place ("numPlaced"). Then, we place the 1st item in the 1st occurrence of that flag, 2nd item in 2nd occurrence, etc. Compare numPlaced to how many we expect to place (Items.Count) for validation.
For this, also reworked how the replacementMap is built to accomodate this - basically, for every location in our game sent by AP, replace it with an item if it's local, or a prism stone if it's not. The rest of the "ItemLot.json" flags (aka known item lot locations) are still added as prism stones, as well, just in case, although the processing also now always adds, for non-AP (but known to us) itemlots, as many prism stones as there are ItemLot object entries for a locid, instead of just 1.
+Also added lots of messaging and validation around this processing.

Added ItemLotParam to easily read in and print out the ItemLotParams from the table we are traversing in memory.
Added back in Archipelago.Core v1.13.
Added some basic try/catch logic to DarkSoulsClient.Connect()
Added a hardcoded switch to turn on text logging at least for the startup+replacement sequence.