Add Overlay messages and Offset Overlay messages below hp + status bar
Added Overlay messages for logger calls that would be made while in game, for Informational, Warning, and Error messages.
Updated the Log.<level> calls to uniformly be Log.Logger.<level>, for consistency.
Set the "y offset" to 250 instead of 100, to be beneath the (default ui scale 100%) DSR health bar and status effects for now. Can be dynamically determined in the future.