Fix up Post-Seath Arena entrance logic
While First Arena after Seath's Death is physically connected to Duke's Archives, imo it more logically comes from Crystal Cave after Seath is defeated. This won't actually effect the logic, because the Event is in Crystal Cave, but I'm simplifying it to make it simpler to read (and to help keep the PopTracker pack up to date).