Add an apworld "api version" to the slotdata
We can use this to validate if the client version is compatible.
Right now it's set to just put error messages out if the third node of the version ("revision") is different. If there's no version, it just puts out warnings (which should be made into errors when that "api" level updates again).
The last node can be updated each time we build, but for for now the 3rd node should be updated each time we make a breaking change on the apworld (e.g. locations change, slotdata becomes incompatible, etc).

There's probably a way to more automatically manage this with the manifest version, but it's easy enough to manage manually for now.

The version also doesn't have to be tied to the client release version - which it would actually be good to be able to separately display to the user.